+++ week04/src/router/router.jsx
@@ -1,0 +1,16 @@
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+
+import Content from '../components/Content';
+import Main from '../pages/Main';
+
+export default function Router() {
+    return (
+        <BrowserRouter>
+            <Routes>
+                <Route path="/search" element={<Main />}>
+                    <Route path=":username" elemnet={<Content />} />
+                </Route>
+            </Routes>
+        </BrowserRouter>
+    );
+}
